Add: Shipment models + endpoints

+++ lib/Shipment.php
@@ -2,7 +2,9 @@
 
 namespace Trunkrs\SDK;
 
+use Trunkrs\SDK\Enum\ShipmentService;
 use Trunkrs\SDK\Exception\GeneralApiException;
+use Trunkrs\SDK\Exception\NotSupportedException;
 use Trunkrs\SDK\Exception\ShipmentNotFoundException;
 
 /**
@@ -12,73 +14,98 @@ class Shipment {
     private static function applyV1(Shipment $shipment, $json) {
         $shipment->id = $json->shipmentId;
         $shipment->trunkrsNr = $json->trunkrsNr;
-        $shipment->labelUrl = $json->label;
+        $shipment->label = new LabelUrls($json);
 
-        $shipment->pickupAddress = new Address($json->sender);
-        $shipment->deliveryAddress = new Address($json->recipient);
+        $shipment->sender = new Address($json->sender);
+        $shipment->recipient = new Address($json->recipient);
         $shipment->timeSlot = new TimeSlot($json->timeSlot);
+    }
+
+    private static function applyV2(Shipment $shipment, $json) {
+        $shipment->trunkrsNr = $json->trunkrsNr;
+        $shipment->sender = new Address($json->sender);
+        $shipment->recipient = new Address($json->recipient);
+
+        array_map(function ($parcelJson) {
+            return new Parcel($parcelJson);
+        }, $json->parcels);
+
+        $shipment->timeSlot = new TimeSlot($json->timeSlot);
+        $shipment->state = new ShipmentState($json->state);
+        $shipment->featureCodes = new FeatureCodes($json->featureCodes);
+        $shipment->service = $json->service;
     }
 
     /**
      * Creates a new shipment for the specified shipment details.
      *
-     * @param ShipmentDetails $details The details of the shipment.
-     * @param Address $pickup The pickup address for the shipment.
-     * @param Address $delivery The delivery address for the shipment.
-     * @param int $timeslotId An optional timeslot, if not specified will be placed in next available.
+     * @param ShipmentDetails $shipment The details of the shipment.
      * @return array The created shipments in an array as instance of Shipment.
      * @throws Exception\NotAuthorizedException When the credentials are invalid, not set or expired.
      * @throws Exception\ServerValidationException When the request payload doesn't match the expectation of the API.
      * @throws Exception\GeneralApiException When the API responds with an unexpected answer.
      */
-    public static function create(
-        ShipmentDetails $details,
-        Address $pickup,
-        Address $delivery,
-        int $timeslotId = -1
-    ): array {
-        $body = [];
-        switch (Settings::$apiVersion){
-            case 1:
-                $body = array_merge(
-                    $pickup->serialize('pickup'),
-                    $delivery->serialize('delivery'),
-                    $details->serialize()
-                );
-                if ($timeslotId !== -1) {
-                    $body['timeSlotId'] = $timeslotId;
-                }
-
-                break;
-        }
-
-        $jsonResult = RequestHandler::post("shipments", $body);
-        if (!is_array($jsonResult)) {
-            $jsonResult = [$jsonResult];
+    public static function create(ShipmentDetails $shipment): array {
+        $results = RequestHandler::post('shipments', $shipment->serialize());
+        if (!is_array($results)) {
+            $results = [$results];
         }
 
         return array_map(function ($result) {
             return new Shipment($result);
-        }, $jsonResult);
+        }, $results);
     }
 
     /**
      * Find the details for the specified shipment by its identifier.
      *
-     * @param int $shipmentId The shipment identifier.
+     * @param string $trunkrsNr The shipment Trunkrs number.
      * @return Shipment A shipment instance.
      * @throws ShipmentNotFoundException When the specified shipment couldn't be found.
      * @throws Exception\NotAuthorizedException When the credentials are invalid, not set or expired.
      * @throws GeneralApiException When the API responds with an unexpected answer.
+     * @throws NotSupportedException When
      */
-    public static function find($shipmentId): Shipment {
+    public static function find(string $trunkrsNr): Shipment {
+        if (Settings::$apiVersion == 1) {
+            throw new NotSupportedException('Please use the shipment id instead of the Trunkrs number.');
+        }
+
         try {
-            $json = RequestHandler::get(sprintf("shipments/%d", $shipmentId));
+            $json = RequestHandler::get(sprintf("shipments/%s", $trunkrsNr));
             return new Shipment($json);
         } catch (GeneralApiException $exception) {
             $isShipmentNotFound = $exception->getStatusCode() == 404;
             if ($isShipmentNotFound)  {
-                throw new ShipmentNotFoundException($shipmentId);
+                throw new ShipmentNotFoundException($trunkrsNr);
+            }
+            throw $exception;
+        }
+    }
+
+    /**
+     * Find the details for the specified shipment by its identifier.
+     *
+     * @deprecated
+     * @param int $id The shipment identifier.
+     * @return Shipment A shipment instance.
+     * @throws ShipmentNotFoundException When the specified shipment couldn't be found.
+     * @throws Exception\NotAuthorizedException When the credentials are invalid, not set or expired.
+     * @throws GeneralApiException When the API responds with an unexpected answer
+     * @throws NotSupportedException When this functionality is called against the V2 API.
+     */
+    public static function findById(int $id): Shipment {
+        if (Settings::$apiVersion == 2) {
+            throw new NotSupportedException('Please use the Trunkrs number instead of the shipment id.');
+        }
+
+        try {
+            $json = RequestHandler::get(sprintf("shipments/%d", $id));
+            return new Shipment($json);
+        } catch (GeneralApiException $exception) {
+            $isShipmentNotFound = $exception->getStatusCode() == 404;
+            if ($isShipmentNotFound)  {
+                throw new ShipmentNotFoundException($id);
             }
             throw $exception;
         }
@@ -101,13 +128,43 @@ class Shipment {
     }
 
     /**
-     * Cancels the shipment by its identifier.
+     * Cancels the shipment by its Trunkrs number. Only available on the V2 API.
      *
+     * @param string $trunkrsNr The shipment id of the shipment to cancel.
+     * @throws Exception\NotAuthorizedException When the credentials are invalid, not set or expired.
+     * @throws Exception\GeneralApiException When the API responds with an unexpected answer.
+     * @throws NotSupportedException When this functionality is called against the V1 API.
+     */
+    public static function cancelByTrunkrsNr(string $trunkrsNr) {
+        if (Settings::$apiVersion == 1) {
+            throw new NotSupportedException('Please use the shipment id instead of the Trunkrs number.');
+        }
+
+        try {
+            RequestHandler::delete(sprintf('shipments/%d', $trunkrsNr));
+        } catch (GeneralApiException $exception) {
+            $isShipmentNotFound = $exception->getStatusCode() == 404;
+            if ($isShipmentNotFound)  {
+                throw new ShipmentNotFoundException($trunkrsNr);
+            }
+            throw $exception;
+        }
+    }
+
+    /**
+     * Cancels the shipment by its identifier. Only available on the V1 API.
+     *
+     * @deprecated
      * @param int $id The shipment id of the shipment to cancel.
      * @throws Exception\NotAuthorizedException When the credentials are invalid, not set or expired.
      * @throws Exception\GeneralApiException When the API responds with an unexpected answer.
+     * @throws Exception\NotSupportedException When this functionality is called against the V2 API.
      */
     public static function cancelById(int $id) {
+        if (Settings::$apiVersion == 2) {
+            throw new NotSupportedException('Please use the Trunkrs number instead of the shipment id.');
+        }
+
         try {
             RequestHandler::delete(sprintf('shipments/%d', $id));
         } catch (GeneralApiException $exception) {
@@ -120,6 +177,7 @@ class Shipment {
     }
 
     /**
+     * @deprecated The shipment id is deprecated in favor of the Trunkrs number.
      * @var int The shipment identifier.
      */
     public $id;
@@ -130,14 +188,14 @@ class Shipment {
     public $trunkrsNr;
 
     /**
-     * @var Address $pickupAddress The pick-up address details for this shipment.
+     * @var Address $sender The pick-up address details for this shipment.
      */
-    public $pickupAddress;
+    public $sender;
 
     /**
-     * @var Address $deliveryAddress The delivery address for this shipment.
+     * @var Address $recipient The delivery address for this shipment.
      */
-    public $deliveryAddress;
+    public $recipient;
 
     /**
      * @var TimeSlot $timeSlot The timeslot in which the shipment has been created.
@@ -145,9 +203,25 @@ class Shipment {
     public $timeSlot;
 
     /**
-     * @var string The label url of shipment.
+     * @var LabelUrls The label urls of the shipment.
      */
-    public $labelUrl;
+    public $label;
+
+    /**
+     * @var ShipmentState The current state of the shipment. Only available as property under the V2 API.
+     */
+    public $state;
+
+    /**
+     * @var FeatureCodes The feature codes set for this shipment.
+     */
+    public $featureCodes;
+
+    /**
+     * @see ShipmentService
+     * @var string The service level requested for this shipment.
+     */
+    public $service = ShipmentService::SAME_DAY;
 
     /**
      * Shipment constructor.
@@ -160,6 +234,9 @@ class Shipment {
                 case 1:
                     self::applyV1($this, $json);
                     break;
+                case 2:
+                    self::applyV2($this, $json);
+                    break;
             }
         }
     }
@@ -170,11 +247,11 @@ class Shipment {
      * @return string The tracking url.
      */
     public function getTrackingUrl(): string {
-        return sprintf("
-            %s/%s/%s",
+        return sprintf(
+            "%s/%s/%s",
             Settings::$trackingBaseUrl,
             $this->trunkrsNr,
-            str_replace(" ", "", $this->deliveryAddress->postal)
+            str_replace(" ", "", $this->recipient->postal)
         );
     }
 
@@ -199,7 +276,7 @@ class Shipment {
         return Label::download(
             $type,
             $this->trunkrsNr,
-            $this->deliveryAddress->postal
+            $this->recipient->postal
         );
     }
 
@@ -209,6 +286,13 @@ class Shipment {
      * @throws Exception\GeneralApiException When the API responds with an unexpected answer.
      */
     public function cancel() {
-        self::cancelById($this->id);
+        switch (Settings::$apiVersion) {
+            case 1:
+                self::cancelById($this->id);
+                break;
+            case 2:
+                self::cancelByTrunkrsNr($this->trunkrsNr);
+                break;
+        }
     }
 }
